# dyninst/object: keep dwarf sections off the heap