### What does this PR do?

This keeps the data off the heap. For uncompressed sections it also keeps
them file backed. This has a somewhat negative impact on decoding throughput
that is a bit surprising. I imagine that it has to do with the faults when
writing the data out, but I'm not certain. The benchmarks definitely are
noisy.

```
name                      old time/op    new time/op    delta
LoadElfFile/cockroach-10     904ms ± 6%     918ms ± 2%     ~     (p=0.067 n=20+18)
LoadElfFile/dd-agent-10      241ms ± 3%     254ms ± 3%   +5.17%  (p=0.000 n=19+19)

name                      old speed      new speed      delta
LoadElfFile/cockroach-10   290MB/s ± 6%   285MB/s ± 3%   -1.69%  (p=0.041 n=20+19)
LoadElfFile/dd-agent-10    297MB/s ± 3%   283MB/s ± 3%   -4.92%  (p=0.000 n=19+19)

name                      old alloc/op   new alloc/op   delta
LoadElfFile/cockroach-10     276MB ± 0%      14MB ± 0%  -95.02%  (p=0.000 n=20+17)
LoadElfFile/dd-agent-10     75.3MB ± 0%     3.6MB ± 0%  -95.27%  (p=0.000 n=19+19)

name                      old allocs/op  new allocs/op  delta
LoadElfFile/cockroach-10      121k ± 0%      121k ± 0%   +0.00%  (p=0.000 n=20+17)
LoadElfFile/dd-agent-10      32.1k ± 0%     32.1k ± 0%     ~     (all equal)
```

Relates to https://datadoghq.atlassian.net/browse/DEBUG-4026